// ClientApp/src/app/shared/models/note.model.ts
export interface Note {
  activityTypeDescription: string;
  employeeDescription: string;
  documentTypeDescription: string;
  customerCode: string;
  noteCounter: number;
  date: string | Date;
  isReserved: boolean;
  activityTypeCode: string;
  appointmentId: string;
  subject: string;
  description: string;
  employeeCode: string;
  documentTypeCode: string;
  documentNumber: string;
  documentDate: string | Date;
  rowGuid: string;
}